added the creation of ets to hold status info for each quarter.
added an external function to call from server to send entire ets to main_pc for back up. this
 function calls handle_cast with the atom periodic_sensor_status_update. the function will update the corresponding ets with the new data(replace it because the ets is a set).

added a sketch of a nodedown event handling: in the message there should be the PID of the node that crashed. if we know it, then we know who to check to give the responsibility to. for example, PC1 crashed, then we check with PC2,PC3,PC4. if PC2 answers, we will send the ets of the pc that has crashed.
there is a scenario for each pc crash.